refactor: refactor code according to review
