+++ src/components/ProductList/index.css
@@ -1,13 +1,11 @@
 .product-list {
   text-align: center;
-  padding: 0 10px;
 }
 
 .products-grid {
   display: grid;
   grid-template-columns: repeat(3, 1fr);
-  gap: 20px;
-  padding: 20px;
+  gap: 54px;
 }
 
 .product-card {
@@ -26,7 +24,7 @@
 
 .product-card img {
   width: 100%;
-  height: auto;
+  height: 380px;
   border-radius: 10px;
 }
 
@@ -46,6 +44,10 @@
   opacity: 0;
   pointer-events: none;
   transition: opacity 0.3s ease;
+}
+
+.icon-overlay img {
+  height: auto;
 }
 
 .image-container:hover .icon-overlay {
@@ -122,6 +124,12 @@
 
   .view-all-container button {
     font-size: 18px;
+  }
+
+  .product-card img {
+    width: 100%;
+    height: 136px;
+    border-radius: 10px;
   }
 }
