+++ src/components/RangeSlider/types/filter.ts
