+++ src/pages/Registration/index.tsx
@@ -7,7 +7,7 @@ import { useForm } from 'react-hook-form'
 import { Link, useNavigate } from 'react-router-dom'
 
 import { auth, registerWithEmailAndPassword } from '../../utils/firebase'
-import { ENschema } from './RegistrationShema'
+import { ENschema } from './validation'
 
 function Registration() {
   const [email, setEmail] = useState('')
