+++ src/store/store.ts
@@ -5,11 +5,9 @@ import priceRangeReducer from './reducers/priceRangeSlice'
 import productsReducer from './reducers/productsSlice'
 import sortReducer from './reducers/sortSlice'
 import themeReducer from './reducers/themeSlice'
-import userReducer from './reducers/userSlice'
 
 const rootReducer = combineReducers({
   themeReducer,
-  userReducer,
   categoryReducer,
   sortReducer,
   productsReducer,
