--- src/types/user.ts
@@ -1,22 +1,0 @@
-export interface IUser {
-  address: {
-    geolocation: {
-      lat: string
-      long: string
-    }
-    city: string
-    street: string
-    number: number
-    zipcode: string
-  }
-  id: number
-  email: string
-  username: string
-  password: string
-  name: {
-    firstname: string
-    lastname: string
-  }
-  phone: string
-  __v: number
-}
