+++ src/utils/firebase.tsx
@@ -21,13 +21,17 @@ const app = initializeApp(firebaseConfig)
 const auth = getAuth(app)
 const db = getFirestore(app)
 
-const logInWithEmailAndPassword = async (email: string, password: string) => {
+const logInWithEmailAndPassword = async (
+  email: string,
+  password: string,
+  setError: (message: string) => void
+) => {
   try {
     await signInWithEmailAndPassword(auth, email, password)
   } catch (err) {
     console.error(err)
     if (err instanceof Error) {
-      alert(err.message)
+      setError(err.message)
     }
   }
 }
